requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.2
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
Django==3.1
django-appconf==1.0.4
django-ckeditor==5.9.0
django-formtools==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-tabler==0.0.32
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jsonfield==3.1.0
Markdown==3.2.2
msgpack==1.0.0
numpy==1.19.1
pandas==1.1.1
pilkit==2.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
Twisted==20.3.0
txaio==20.4.1
xlrd==1.2.0
zope.interface==5.1.0
